Correct the error levels to hide non severe errors.
This is needed as the notices from the legacy code are confusing (at
least) the console application making it unusable.
Using this we have also dropped support for defining the error levels
via localconfig.php etc. as it makes the system unstable when legacy
code is active.
I can not think of a proper way to keep this feature and at the same
time maintain a stable system aside from fixing all notices.