change isset to array_key_exists
isset returns "false" if the array key value is NULL. This is an issue by updating fields with value NULL, because they don't get marked as modified.
At this point it is only necessary to check if this key is present in those arrays. The value doesn't matter.

Bonus: No NOTICE logs for undefined array index.