Handle exceptions in the initialize.php entry point (see #525)
Description
-----------

On custom entry points (using `initialize.php`), `ResponseException`s are not caught and not converted to a response. This attempts to fix this (and also automatically adds debug output etc. if `kernel.debug` is enabled.

see https://github.com/isotope/core/issues/1993

Be aware that some code needs to be updated when merging this into master!

Commits
-------

9907e70a Handle exceptions in the initialize.php
ba2976de Fix the coding style