Added support for configuring the SearchIndexListener behaviour (see #983)
Description
-----------

This will allow to configure how the `SearchIndexListener` behaves.
Right now it updates the index when the response was successful and it deletes it from the index, if it was not.

If you have cronjobs doing that work for you, you may want to disable that listener completely or have it e.g. only delete from the index. You can configure all 4 possible states using the config now:

```yaml
contao:
    search:
        listener:
            index: true
            delete: false
```

Commits
-------

d31c0495 Added support for configuring the SearchIndexListener behaviour
792a6bad Remove all search indexers including the delegating one and disable the searchindexlistener if none has been given
cf3de3b3 Made configuration of the search index listener easier
3fcde709 Merge branch 'master' into feature/searchindexlistener-configuration
2c3ed6fc Fix the coding style