Rework the front end preview (see #989)
Description
-----------

This PR injects the back end preview bar and replaces the preview frameset.

**How this works:**

The preview bar is injected into the DOM on `kernel.response` event.
When the user wants to view the page as a member, an Ajax request is triggered in which the preview authentication will be handled. The Ajax endpoint logs in the user, returns an HTTP OK  and finally, a page reload is done.

**TODO:**
- [x] CSS (rudimental CSS at the moment)
- [x] make the preview bar collapsible
- [x] restore contao/core#6546

**Impressions:**
<img width="1411" alt="Screen Shot 2019-11-24 at 16 09 40" src="https://user-images.githubusercontent.com/1284725/69491274-38fb3980-0ed6-11ea-97b6-625b730c9505.png">

____

**Outlook:**
I already worked on an enhanced version of the preview. Once this PR is approved and merged, I add a framework for 3rd parties to add content to the preview toolbar.
So-called "preview providers" (class must implementing an interface and tag the service with `contao.preview_provider`) can add arbitrary content to the toolbar.
In the end, the great advantage of the preview is the back end scope in the front end.
The toolbar should be either horizontally (as we know it) or left-aligned when the content requires it.

Use case: theme switch or changing CSS variables on runtime.

<img width="1509" alt="Screen Shot 2019-11-24 at 20 52 19" src="https://user-images.githubusercontent.com/1284725/69494228-57742b80-0efc-11ea-9254-0a1848112b65.png">

Commits
-------

414b966a Rework frontend preview
4bad3d62 Collapse toolbar
2b856c1c Check for frontend request
a3a56303 Apply suggestions from code review
2098eba0 Remove the contao_backend_switch route
8eacd71f Merge remote-tracking branch 'richardhj/feature/preview' into feature/preview
3315cc90 Apply suggestions from code review
b04c925b Inject js before </body>
9ebe83d5 Remove non-present route
6fae43a8 Move logic to PreviewUrlConvertEvent
280d16a2 Add tests
2336e22f Apply suggestions from code review
2974a205 Do not inject preview bar if no preview entrypoint defined
dfe0d2a9 Add ROLE_ALLOWED_TO_SWITCH_MEMBER
519124dc Further styling of toolbar
07327656 Fix safari incompatibilites
cfd83d93 Merge branch 'master' into feature/preview
80d06fda Fix QS
f2f36f4a Fix the coding style
0e3ae2e3 Merge remote-tracking branch 'upstream/master' into feature/preview
4a969b25 Serialize amg in BackendUser.php
206d929b Remove CDATA as we use HTML
2c1c97d9 Do not initialize framework.
550a5612 Test service registration.
836d3c3a Fix the coding style
1371041b Fix the CSS code
8cbf0169 Simplify event listener registration.
7666ac6a Remove legacy preview script code.
b225e1d2 Add BackendPreviewSwitchControllerTest.php
973a9794 Adjust the test method names