Correctly handle URL suffix when redirecting page IDs (see #1503)
Description
-----------

I'm not entirely sure this patch is correct. Does `Environment::get('relativeRequest')` contain the query parameter, or does it not? Depending on that, the `$` should be added on line 186, but it sure should be removed on line 184 imho.

Commits
-------

0e0e69d0 Correctly handle URL suffix when redirecting page IDs
743d1275 Remove the end delimiter in regex
bca80192 Correctly match end of suffix
b3a053d0 Let preg_quote() check for an empty string