Correctly sort if both root pages are fallback (see #1533)
Description
-----------

While working on #1516 I discovered a bug in the route sorting algorithm. In the current functional tests, we have the following case:

- page 1  with alias `main` in domain `A` (which is fallback)
- page 2 with alias `main` in domain `B` (which is fallback)
- page 3 with alias `main/sub` in domain `B` (which is fallback)

Fetching a route for `main/sub` did not correctly return page 3, because page 2 was sorted (and matched) before it.
Because the database/candidates algorithm sorted the database results by length of alias, the tests didn't fail yet. But that sorting shouldn't be necessary. If the database items are not fetched by that order, the sorting algorithm must sort accordingly.

Commits
-------

b6e66470 Correctly sort if both root pages are fallback
5e905bd8 Added unit test
b4cb2ce4 CS
78de722c CS