Allow clearing the model registry (see #1182)
Description
-----------

This PR adds an option to clear the model registry.

This would be useful for functional tests that access Contao Mondels for instance. Here clearing the registry after each test class is essential for the isolation of test cases. Currently this can only be done by calling refresh on each model instance (which is unpractical) or by using reflection to access the registry's fields (which is hacky).

Commits
-------

7dd117a6 allow clearing the model registry
b8ef6dc6 rename method to reset
38000c70 add note about internal usage
0ff32c4b call reset on various legacy classes during kernel.reset
e27079a4 harden array access if key does not exist
6ce4363c test if reset calls are delegated
55985b6d Update core-bundle/src/Resources/contao/library/Contao/Model/Registry.php

Co-Authored-By: Leo Feyer <github@contao.org>
8f5fff7b fix unit-tests on ci
00f33f54 use page model
3f1a6828 CS