Correctly execute kernel events for initialize.php (see #1410)
Description
-----------

Yet another fix for the legacy `initialize.php`

In https://github.com/isotope/core/issues/2064 we are modifying the session to store a state. Unfortunately, the session is not stored in our current approach, because that's done inside a `kernel.response` listener.

This hacky approach fixes this. It catches all the output of the entry point script, generates a response from it and sends it to the client.

**TODO:**
- [x] Check if we need to handle `HTTP/1.1 xxx` status code headers

Commits
-------

b8c5951f Correctly execute kernel.response and kernel.finish_request for initialize controller
6e256be9 Correctly handle status code header
2e024e3f Make sure HTTP status code header is merged in the MergeHttpHeadersListener
c1e1435d CS
eeb3a19b CS and test fixes
1d7b7339 Fix missing services
2a3325bd Drop the InitializeControllerTest
5462df8c Added listener to stop kernel.response event for the InitializeControllerResponse
ff09091a Make the output buffer removeable
33c11175 Fix a phpDoc error
66575d8c Correctly match the HTTP header