Increase the split button breakpoint (see #1608)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1484
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

cf1a7f25 Increase the split button breakpoint