Verify TOTP with a window of 1 for better UX (see #1634)
Description
-----------

I think we should allow for a TOTP window of 1 when checking the validity of a code.
I often find myself not being able to log in because the time the code was copied into my clipboard by my password manager, the token was only valid for another second or two. Until I hit enter, the code has already become invalid and I cannot log in.
By allowing a window of 1, the previous code remains valid. So the time a code is considered valid increases from 30 seconds to 60 seconds (at max) but it increases UX considerably and is thus done in many applications (Google authenticator does this too for example).

Commits
-------

18e64832 Verify TOTP with a window of 1 for better UX