Allow to strip query parameters in reverse proxy (see #1624)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | TODO once merged

This PR will allow to strip query parameters that are not required for the application but only for e.g. some tracking tools in the front end. It increases the cache hit rate and prevents these parameters from flooding the cache.

While implementing this I thought there's also the use case where people trust the Contao developers to maintain a proper blacklist but only need one exception to that. Like e.g. you need the `utm_*` parameters in PHP but you don't want to whitelist all of the allowed query parameters.
Same might apply for cookies so I unified that :)

Commits
-------

2c0ad692 Allow to strip query parameters in reverse proxy
1d957d0a Merge branch 'master' into feature/strip-query-params-subscriber