Skip the preview redirect if the preview script is not set (see #1692)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1511
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

a80b39ca Skip the preview redirect if the preview script is not set
31f08fd3 Merge branch '4.9' into fix/preview-redirect