Improve performance of file manager (see #1839)
Description
-----------

I found a huge performance bottleneck in the file manager.

To check for permissions, we count all the files for every single delete button callback recursively. However, we only need to know if there is at least something so the actual number is completely irrelevant to us and we can early return.

Saved 2 seconds of loading time for my use case :)

Commits
-------

e0f0b66e Do not count all the files recursively, we only need to know if there is at least one result so we can early abort