Remove redirect status type from 401 and 403 (see #1762)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | N/A
| Docs PR or issue | N/A

## Description

Lately the following issue has come up more frequently in the community:

1. Create a login module with __Redirect to last page visited__ enabled.
2. Create a login page and insert the login module.
3. Create a 401 page, enable __Forward to another page__ and select the login page as the __Redirect page__. _Leave_ __Redirect type__ _at its default value!_
4. Create a protected page.
5. Open the protected page in the front end - you will be redirected to the login page.
6. Log in as a front end member.
7. You will "stay" on the login page after logging in, instead of being redirected to the previously visited protected page.

What happens is this:

1. Client requests a protected page.
2. Server answers with a `301` redirect to the log in page.
3. Client logs in.
4. Server answers with a `302` redirect to the previous page, i.e. the initially requested protected page.
5. Client reloads the log in page, since the `301` redirect for the protected page to the log in page is still cached by the client.

I think it does not make sense to be able to select a redirect type for the 401 and 403 pages, as you'd never want a `301` redirect, only some sort of temporary redirect. This PR removes the `redirect` field from the `autoforward` subpalette for the `page_401` and `page_403` page types and redirects with the default `303` status code instead.

## Questions

### Status Code

Which redirect status is correct in this case, `303` or `307`? I am never sure with those.

### Callback Service(s)

Since I am using a tagged service for the `config.onload` callback, I am wondering how such services should be structured and organized within the CoreBundle now and in the future. I chose a general `PageCallbackListener` class, with the idea, that any `tl_page` callbacks are collected there in the future, in individual methods, mirroring the current `tl_page` class. 

However, if we hypothetically added all the current callbacks currently used in the `tl_page` DCA to that class, it would get kind of bloated. So alternatively we could also use small, invokable services for each individual callback task for example. Or structure the classes by callback type (`config.onload`, `config.oncut`, `fields.….save`, …).

## Contao 4.4

If we agree that this is a bugfix, I'd make a separate PR for the `page_403` page type in Contao 4.4.

Commits
-------

3b35274e remove redirect status type from 401 and 403
18a039fe move listener to new namespace
9e58f752 alphabetize listener.yml
835a8765 alphabetize listener.yml
8528aa6e do not use empty for ->id
2eed7618 remove Page sub namespace
5060ba73 use an early out
920147a4 remove redirect status from autoforward subpalette in general