Fix metadata wording (see #2054)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes none
| Docs PR or issue | https://github.com/contao/docs/pull/502

It seems we've got the english word "metadata" wrong at several places. It's a single word (not "meta data") - see https://en.wikipedia.org/wiki/Metadata for instance. Although it appears as a two word variant at some places, it's much more commen as a single word. Also Symfony and Doctrine use this (e.g. `MetadataAwareInterface`, `ClassMetadata`, ...).

Most of this was introduced with the Image Studio PR :see_no_evil: which we could still fix before releasing 4.10. 

This PR now adjusts it everywhere with the exception of the  `Frontend::getMetaData` function signature that existed before 4.10 and cannot be changed for BC reasons.

Commits
-------

46307bd2 rename "meta data" -> "metadata"
2692f080 revert ld+json PageMetaData (see https://github.com/contao/contao/pull/2054#discussion_r463978795)