Allow to not redirect to a language (see #2017)
Description
-----------

Using this new flag, one can configure a root page to not redirect to the language. Be aware that this is **not** the same as `doNotRedirectEmpty`, quite the contrary. 

It is now possible to define a root page with empty URL prefix, and several others with URL prefix. The user will always see the "empty prefix" page without being redirected when enabling the checkbox on all root pages.

Commits
-------

4ebcbba0 Allow to not redirect to a language
c50e0ff4 Add PageModel method hints
343657f0 Merge branch '4.10' into feature/no-language-redirect
e2704955 Rename "doNotRedirectLanguage" to "disableLanguageRedirect"