Use HTTP status code 303 instead of 307 for redirects (see #2096)
Description
-----------

A 307 redirect does not change the request method and therefore breaks the Contao login (and other routines that are triggered with a POST request and expect the following redirect to be a GET request).

The PR reverts the changes from #2037 but replaces 302 with 303 where applicable.

Commits
-------

e64c4297 Use HTTP status code 303 instead of 307 for redirects
dad4352b Use the Symfony default status code in the redirect response