Also add template options to legacy template data (see #2071)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

The `FigureBuilder` allows defining template options that are shipped to the template without any processing. This can already be used in the Twig templates to add classes but should also be present in the legacy template data if you want to use the functionality in your own templates.

As a side-effect this enables overwriting any of the result properties before they get applied to the template which is helpful to handle some special/BC cases (e.g. overwriting the `linkTitle` instead of using the semantic link attribute `title`).

Commits
-------

7bb0dbfc also merge template options into legacy template data