Add custom exception to further handle route parameter errors (see #2019)
Description
-----------

This is an idea to handle routing issues. It would now be possible to handle missing parameters (e.g. ask the user). I'll try to follow up with a concept for the front end preview.

Commits
-------

bca923bb Render an exception when URL parameters are missing
dc7b4404 CS
35a11bc9 Only show the exception in the back end
4fa055c0 Use language files
01e5876f Add CSS for the table
3455c3e2 Update core-bundle/src/Resources/contao/languages/en/exception.xlf

Co-authored-by: Leo Feyer <github@contao.org>
83a33c03 Update core-bundle/src/Resources/contao/languages/en/exception.xlf

Co-authored-by: Leo Feyer <github@contao.org>