$this->ptable not available in the DataContainer class (see #2103)
Description
-----------

In #2039, we have changed the `invalidateCacheTags()` method, which now accesses `$this->ptable`. However, `protected $ptable` is only declared in the `DC_Table` class and not in the `DataContainer` class.

Commits
-------

d2a8183d Move the $ptable class property to the DataContainer class
76a01b15 Use the DCA array instead of $this->pid
115ecca2 Move the method to the DC_Table class
d55f2c73 Revert moving the method