Support captcha input wrapped in DIV (see #2155)
Description
-----------

This is just a quick fix for my use case but it can't really harm. On my client's system, the input element was wrapped in a `div`, therefore the script didn't hide the label. I didn't want to rewrite the whole script and implement recursive DOM traversal, but this adjustment might fix some edge cases.

Commits
-------

0a09e9bf Support captcha input wrapped in DIV