Work towards psalm level 6 (part 2) (see #2173)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

This is a follow-up to #2124 that addresses the issues reported in the `Contao/Image` namespace.

After playing around with some solutions for the `Figure` class I think it's best/most readable to just annotate the return types. This way we can keep the `resolveIfClosure` method which imo helps devs to understand what the code is doing and also makes sure we're evaluating each closure with the same argument.

Commits
-------

31ec94d5 add tuple annotation + remove null from possible values
fa4b9c3d help psalm understand the evaluated type

(it might be able to infer this from the condition in the called hassers in the future but currently cannot)
06630ab6 remove redundant return statement