Do not change CSRF token cookie, if response is not successful (see #2252)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2246
| Docs PR or issue | -

I am not sure if this is the right approach to fix #2246, but I think it makes sense to never change the CSRF token cookie, if the response is not within the `>= 200, < 300` status code range. Implementing this check would fix #2246 as far as I tested.

The problem is, that an AJAX request on the same page that requires the presence of a CSRF token cookie can invalidate said CSRF token cookie, if the generated response in turn requires a new CSRF token cookie (e.g., if the 404 page redirects to another Contao page which in turn requires a CSRF token cookie).

The problem does not occur in Contao 4.4, at least not with the reproduction I posted.

Commits
-------

a7c17cf6 do not change CSRF token cookie, if response is not succesful
59271940 Merge remote-tracking branch 'composer/4.9' into fix-csrf-token-cookie-deletion
04934527 change isSuccessful check
3880e354 Update core-bundle/src/EventListener/CsrfTokenCookieSubscriber.php

Co-authored-by: Martin Auswöger <martin@auswoeger.com>
10cc70fd Update core-bundle/src/EventListener/CsrfTokenCookieSubscriber.php

Co-authored-by: Leo Feyer <github@contao.org>