Consider symlink when saving localconfig.php (see #2209)
Description
-----------

We are using [Deployer](https://deployer.org/) to deploy our Contao 4 instances. In the config, we have added

```
system/config/localconfig.php
``` 

to the `shared_files` so that any changes to the system config in the back end do not get lost after the next deployment. However, this does not actually work as Contao simply overwrites the symlink with a regular `localconfig.php` file when saving.

This PR changes that: it evaluates the `realpath` of an existing `localconfig.php` and uses that path as the new destination.

Commits
-------

84c34b8b consider symlinks when saving localconfig.php
3d6582b6 Merge branch '4.4' into consider-symlinks-for-localconfig
5a895ca5 remove unnecessary file_exists call

`realpath` already checks for the existence of the file and returns `FALSE` in that case.
5489f4c5 fix typo
d828189b add PR reference
99155c62 Remove PathUtil
14a67e5f Re-add PathUtil