Correctly detect unknown options (see #2360)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2349
| Docs PR or issue | -

This fixes all the issues mentioned in #2349. However, there is still a problem that I could not yet solve:

<img width="598" alt="" src="https://user-images.githubusercontent.com/1192057/94276257-45ae4b00-ff48-11ea-87e6-a35083150ec9.png">

This is because the value of the `ImageSize` widget can be an array with three items (e.g. `[300, 300, 'crop']`) and all of the values are added as unknown options. The `Widget` class does not know that only the third value needs to be checked.

@contao/developers Any idea how to fix this without a protected method that can be overwritten in each widget class?

Commits
-------

e95d1c6c Correctly detect unknown options
0272778d Pass the unknown options as attributes
bf962dba Add more widgets
1a4cbf63 Add an is_array() check before the foreach loop