Fix SimpleTokenParser BC (see #2341)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2315, #2387 
| Docs PR or issue | -

To regain BC there now is an additional step in which the variables used in the expression are identified. With this we can find out if any variable isn't provided in the array of data and log + exit gracefully (= return `false`) instead of throwing an exception. This also allows logging all unknown tokens not only the first occurrence.

I also added a `bool` cast to the evaluated result to support shorthand/non-boolean expressions like `{if foo}`, `{if 1}`, `{if a + b}`.

Commits
-------

321aebfa setup some BC tests
1ee74ae2 correctly handle unmatched tokens
5e45ff7f add and adjust tests
193723fc drop BC test dummies
ec3d58dc preset missing vars with 'null' instead of always returning false
d7995b68 also treat missing var as 'null' in legacy code path