Add createIfDeferred to ImageResult (see #2368)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Closes #2322
| Docs PR or issue | Todo

This adds the ability to synchronously trigger the creation of resized images from within an `ImageResult`. This is similar to `File#createIfDeferred()` but processes all resources (picture img + picture sources) instead of a single file.

Contrary to the method in `File` an exception will be thrown if the `contao.image.resizer` service does not support deferred resizing (`DeferredResizerInterface`) and exceptions won't be swallowed.

**Demo Usage:**
```php
$figure = $studio->createFigureBuilder()
    ->fromPath('files/public/foo.jpg')
    ->setSize('_my_size')
    ->build();

$figure->getImage()->createIfDeferred();

// It's now safe to access any of the generated image files directly
$pdf->addImage($figure);
```

Commits
-------

5ffb1945 add contao.image.resizer to studio locator
45815c66 add createIfDeferred method to ImageResult
9c02354f adjust ContaoCoreExtensionTest
71ffe3fc also scan for 'srcset' elements (+ remove duplicates)
b73ba710 do not throw exception if there are no deferred images
4f30e2f9 simplify getting candidates + remove duplicate check
6f0b9765 CS
412ce3ea Merge branch 'master' into feature/studio-create-if-deferred