Fix unknown option when creating or editing error pages (see #2419)
Description
-----------

This fixes a bug introduced in #2397. If you create a 401, 403 or 404 page, it will be shown as an "unknown option" in the type selection.

This is because the `FilterPageTypeListener` is supposed to remove any error pages that are already present within the same website root. However, it also removes the page type of the page being currently edited.

This PR adds an additional query condition to prevent that.

Commits
-------

8fe8b2c6 fix unknown option when editing error pages
28cb8f94 fix unit tests and code style
c418d726 remove dump