Use a listener to reset custom templates (see #2669)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2666

I'm sorry for the late response about #2666, @fritzmg is right there. The behaviour of adding new methods to `tl_` classes is discouraged and something we agreed not to do. Doing the same in a listener is much easier and more universal, see this PR changes. It also means we don't need to duplicate the code in multiple places.

Commits
-------

80a6f6a5 Use a listener to reset custom templates
52824128 Register the listener
4505c323 CS
626922da Add the missing `['config]` part