Correctly handle trailing slashes in URLs (see #2690)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2493

Symfony has a special handling for trailing slashes in URLs.

If the request [matcher implements RedirectableUrlMatcherInterface](https://github.com/symfony/symfony/blob/5.x/src/Symfony/Component/Routing/Matcher/UrlMatcher.php#L135), the regex to match [the URL is modified](https://github.com/symfony/symfony/blob/5.x/src/Symfony/Component/Routing/Matcher/UrlMatcher.php#L151) to allow for URLs with or without trailing slashes.

This is intentional, as the `RedirectableUrlMatcher` will redirect to URLs with slash if one is missing. We use the `RedirectableUrlMatcher` to redirect https to http (and vice versa) as well.

The modification makes the parameter lazy, so trailing slashes are matched even if optional. When matching `/en/news-detail/new-contao-versions-in-short-intervalls/`, parameters must be matched as `new-contao-versions-in-short-intervalls` instead of `new-contao-versions-in-short-intervalls/`

Commits
-------

9c15ad70 Correctly match the URL parameters