Use the chained router to find root pages (see #2717)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2499

I had this issue in a client application, debugging it showed that the fallback 404 was actually found correctly, but `Frontend::getRootPageFromUrl` used by  `PageError404` did not resolve correctly. Most likely because the error404 pages are not present in the regular matcher but only in the chained router/matcher.

Commits
-------

4c84ef97 Use router to find root page
5cebd566 Merge branch '4.9' into bugfix/fallback-404
1015a077 Correctly match 404 page and unknown languages
f05ddf1c Fixed config test
4f6919e0 CS