Sort root page routes after other page routes (see #2819)
Description
-----------

@netzarbeiter and me found another routing edge case similar to https://github.com/contao/contao/pull/2465

In this case, the requested domain only exists in a language the user does not have it her preferred browser languages. In this particular setup, this resulted in incorrect sorting, where the root page was sorted before the page with `index` alias, which caused an endless redirect loop.

Because sorting always depends on a combination of "unknown" database records, I've created a new SQL dump that contains a anonymised copy of the client system.

The fix is to always sort `root` routes after any other, because they are always just a fallback match for when there's no page with `index` alias.

Commits
-------

7757f5b3 Added test case to show failing CI
08196ee9 Fix sorting issue and rename file to pull request ID