Handle image URLs with a {{file::*}} insert tag in the lightbox (see #2674)
Description
-----------

By default `Controller::addImageToTemplate` acts in such a way that if you provided an `imageUrl` with a local image and enabled `fullsize` then the `lightbox` attribute is automatically added, so that the target image is displayed in the lightbox and not in a new browser window/tab.

For example:

1. Create a new image content element.
2. Enable _Overwrite meta_.
3. Set the image URL to `files/foobar/some-image.jpg`.
4. Enable _Full size/new window_.

This will open the image under the provided image URL in the lightbox.

**However**, if you instead do the following:

1. Create a new image content element.
2. Enable _Overwrite meta_.
3. Select the same file from the previous example, but this time via the universal picker in the file manager.
4. Enable _Full size/new window_.

then the same image from before will only open in a new tab/window, instead of in a lightbox.

The reason is that the logic in `Controller::addImageToTemplate` checks whether the provided image URL is a local image - but in the latter example the provided image URL is a `{{file::*}}` insert tag and thus it is not detected a a local image.

This PR fixes that by replacing the insert tag beforehand.

Commits
-------

6c773730 fix lightbox not working when using {{file::*}}
e57d2b11 only handle file insert tags
81965710 Merge remote-tracking branch 'origin/4.9' into fix-lightbox-with-insert-tag-url
c2afcd46 always replace insert tags and only url decode for picture factory
fe3df76e use replaced insert tag when local processing failed