Add a PHPUnit extension to test for deprecated classes (see #2863)
Description
-----------

This is an alternative to #2517

It adds a before-all-tests and an after-all-tests hook to phpunit and tests the deprecated classes before and throws an error after the tests. This has several benefits:

- Test execution order (or using `--filter`) is no longer a problem as all deprecated classes get loaded (and verified) before the first test is executed
- Fixes deprecations for functional tests too, where other methods (like #2517) have problems.
- Single point in the codebase where we can check if we need to actually react to a deprecation when migrating to a new Symfony version for example.

For clarification: This is for file scoped deprecations (= `trigger_error` outside the class definition) that get triggerd only once when the class file gets autoloaded.

Commits
-------

8753ed35 Test deprecated classes in phpunit extension
b29c222c Expect deprecations based on feature detection
42e61820 Move phpunit extension to contao/test-case