Use eval.context in the picker if set (see #2837)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes https://github.com/madeyourday/contao-rocksolid-custom-elements/issues/129

For dynamic DCA fields (like in multicolumn wizard, custom elements and so on) there is no relation information stored in the DCA extractor. The picker widget supports setting `eval.context` to something like `dc.tl_content`, but when rendering the values this `context` is ignored. This PR changes that so that `context` is always used and the relation data as fallback.

Commits
-------

6c44cb50 Use context in picker if set and relations as fallback