Fix loading database.sql files in the DcaExtractor (see #2762)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

The `DcaExtractor` uses this check (that is run on each field of a DCA) to check if a `database.sql` file should be parsed:

https://github.com/contao/contao/blob/712de2476dc6100cad1876da1f22b809f7558d5c/core-bundle/src/Resources/contao/library/Contao/DcaExtractor.php#L400-L404

However, this check is always true in case of an entity backed DCA or if there are any fields without a SQL definition like explanations or fields that are populated/stored by a `load/save_callback`.  We're then falsely triggering a deprecation notice (L439) and an invalid array access occurs (L462).

This PR always tries to locate the `database.sql` files and only continues processing if there are actually any.

Commits
-------

c637f52e do not rely on DCA sql definition to determine if a database.sql file exists
0651990b optimize file access
fa7e27cd adjust wording
02e30fe8 revert old logic
cf7d7e1c Merge branch '4.9' into bugfix/dcaextractor-from-file