Add JSON+LD meta data to the search index (see #2801)
Description
-----------

Closes https://github.com/contao/contao/issues/2738.
The actual part of just extracting the meta data and storing it is super easy but I thought we also need a use case for it.

One thing I keep adding for our customers is displaying an image to the search results. I've checked schema.org if there's a standardized way as to how you can declare a primary image for a webpage and turns out there is: https://schema.org/primaryImageOfPage

So I've implemented that by default.
How that information is added to the page is a different story and not part of this PR.

Commits
-------

cc173497 Added JSON+LD data to search index and use it to display the primary image of a webpage
56eed96f Cleanup
51ad8e3b Merge branch '4.x' into feature/metadata-search

* 4.x: (39 commits)
  Make the FigureBuilder more fail tolerant (see #2733)
  Enable accessing Figure in all templates (see #2734)
  Quote the "group" field in the UserCreateCommand statement (see #2858)
  Add support for namespaced DC drivers (see #2706)
  Always show all errors in the contao-setup binary (see #2845)
  Fix another illegal array access in System::getReferer() (see #2843)
  Add a PHPUnit extension to test for deprecated classes (see #2870)
  Fix the search query if there are no keywords (see #2856)
  Update the changelog and the language files (see #2841)
  Initialize the Contao framework in the Version480Update class (see #2836)
  Adjust the description of the YouTube "rel" option (see #2838)
  Handle null arguments in the ContentCompositionListener (see #2785)
  Fix an illegal object access in the Versions class (see #2835)
  Use dependency injection for the InitializeController (see #2833)
  Allow passing an array of IDs to User::isMemberOf() (see #2834)
  Fix generating error pages by name (see #2827)
  Start the session to check for a user token (see #2829)
  Do not tag contao.db.tl_module.<id> in articles (see #2815)
  Fix an illegal array access in DC_Table when expanding the tree (see #2805)
  Fix the logout handler in Symfony 5 (see #2818)
  ...
f58b1460 Do not use trait anymore
76579abe Pass on figure
cddaedd9 Use the first schema.org appearance
8b9b4402 CS
d186b076 Fixed indexer tests
6ef98ebf Update core-bundle/src/Resources/contao/modules/ModuleSearch.php

Co-authored-by: Martin Auswöger <martin@auswoeger.com>
49ef4f67 Improved template
5248b825 Update core-bundle/src/Resources/contao/templates/search/search_default.html5

Co-authored-by: Martin Auswöger <martin@auswoeger.com>
9d8bd2be CS
01951e7b CS
6480ac29 Move the image size into its own palette
d1da0856 Update core-bundle/src/Resources/contao/modules/ModuleSearch.php

Co-authored-by: Leo Feyer <github@contao.org>