Fix label callback for tree view in picker widget (see #2906)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #2882 

Fix the call of the label callbacks of tree views in the picker widget.
With the `isProtected` boolean I was not sure, if we need to solve that recursive.

Commits
-------

c3c1e9e9 Fix label callback for tree view in picker widget #2882
d8a4d015 Set protected flag for label callback always null