Allow saving the value 0 in the KeyValueWizard (see #3007)
Description
-----------

Replaces the pull request #3006.

---

This PR fixes the bug that 0 values cannot be entered in the KeyValueWizard.

It seems that this possibility was removed in the preparation for PHP8.
In further versions of contao 4.9 it was possible to set 0 as value.

(#2347, #2342)

Commits
-------

e99980e4 Allows to save the value 0 in the keyValueWizard (4.9)