JSON-LD support for Twig (see #3170)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | see https://github.com/contao/contao/pull/3163#issuecomment-875930255
| Docs PR or issue | todo

This refactors the image studio macros to use our new JSON-LD system instead of item* attributes (+ some housekeeping).

I registered the `addSchemaOrg` Twig function in the `ContaoTemplateExtension` extension for now. Note, that this is only temporary. Once #2988 is merged, I'd like to move everything in a single `ContaoExtension`. Having multiple runtimes is a good idea for SOC, but it turned out that multiple extensions doesn't make any sense for our use case (IMHO BC is not relevant in this case).

/cc @Toflar @ausi

Commits
-------

79dcba6d remove item* attributes (+ better whitespace control)
59ff764e add typehint
726d915c auto add schema.org data when rendering a figure
4c060f9a add SchemaOrgRuntime
175f0912 add + adjust tests
637b5599 use snake_case for Twig functions
917e77ba CS