Simplify Controller::addImageToTemplate tests (see #3182)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | -
| Docs PR or issue | -

I initially wrote the `FigureBuilderIntegrationTest`s to make sure the new version of `Controller::addImageToTemplate` (using the `FigureBuilder` internally) behaves identical to the original implementation. The tests were valuable back then but they encapsulate an insane amount of complexity like setting up the legacy framework, hacking the model registry and so on. 

While working on an update to #2518 I realized there are a lot of complication with these tests, therefore this PR.  :slightly_smiling_face: They are now replaced with a much simpler version that mocks things instead.

Commits
-------

7dc3f9e3 replace figure builder integration tests with mocked alternative
072bfa10 cast floored value to int
e3205323 CS
8e0dc9a2 CS